chore(warnings): clear fetch compile graph + Mojo 0.26.2 modernization
Three categories silenced across the libraries fetch imports:

- `@parameter if PROFILE_ACCEPT:` → `comptime if PROFILE_ACCEPT:`
  across src/quic/connection.mojo (27), src/h3/connection.mojo (17),
  src/h3/h3_handler_server.mojo (4), tests/test_quic_profile.mojo (1).
  The new `comptime if` is stricter about dead initializers; added
  `comptime if not PROFILE_ACCEPT: _ = t_start_*` discards in 4 sites
  in src/h3/connection.mojo so the init `= 0` isn't flagged on the
  default (PROFILE_ACCEPT=False) build.

- Implicit stdlib imports → fully qualified `std.*` in
  src/quic/profile.mojo, src/h3/h3_udp_server.mojo,
  tests/test_cc_pacing.mojo, tests/test_h3_udp_server.mojo,
  tests/test_quic_profile.mojo.

- Implicit `Int → UInt32/UInt64` conversions at call sites:
  src/h2/h2_session.mojo (stream_id cast), src/h3/h3_session.mojo
  (stream_id cast).

Also removes a dead `has_initial = True` and renames two unused
`for i in range(_AEAD_TAG_LEN)` loop vars in src/quic/connection.mojo.

The default fetch compile graph (PROFILE_ACCEPT=False) is now
warning-free.

+++ src/h2/h2_session.mojo
@@ -281,7 +281,7 @@ struct H2Session(Session):
         var hid = Int(handle_id)
         if hid not in self._handle_to_stream:
             raise Error("H2Session.feed_body: unknown handle")
-        var stream_id = self._handle_to_stream[hid]
+        var stream_id = UInt32(self._handle_to_stream[hid])
         if frame.is_data():
             var bytes_copy = frame.data().copy()
             self._conn.send_data(stream_id, bytes_copy^, end_stream=False)
